% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bls_download}
\alias{bls_download}
\title{Download data from the Bureau of Labor Statistics}
\usage{
bls_download(seriesid, start_year, end_year, bls_key)
}
\arguments{
\item{seriesid}{Character vector. Vector of series IDs for BLS data. Limit of 50 strings.}

\item{start_year}{Numeric.}

\item{end_year}{Numeric.}

\item{bls_key}{Character. BLS API key. Required parameter.}
}
\description{
\code{bls_downloads} remotely downloads data for a given set of series IDs from the Bureau of
Labor Statistics (BLS).
}
\details{
This function acts as a wrapper for the function \code{bls_api} from the package
\code{blscrapeR}. Please see documentation for \code{bls_api}.
}
\examples{
# National series for seasonally adjusted private sector employment and earnings
series_str = ces_seriesid(adjustment = c("S"), industries = c("05000000"),
data_types = c("01", "03"))

bls_df = bls_download(seriesid = series_str, start_year = 2006, end_year = 2010,
bls_key = Sys.getenv("BLS_KEY"))
}
