% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces.R
\name{ces_emp}
\alias{ces_emp}
\title{Download national employment data from Current Employment Statistics (CES) survey}
\usage{
ces_emp(bls_key, series = "nfp", start_year, end_year, adjustment = "S")
}
\arguments{
\item{bls_key}{BLS API key for the user. See vignette.}

\item{series}{Character. Data series. Either non-farm payrolls ("nfp"), private payrolls
("private"), supersector payrolls ("super"), or sector payrolls ("sector"). Defaults to non-farm
payrolls ("nfp").}

\item{start_year}{Numeric. Year to start data download.}

\item{end_year}{Numeric. Year to end data download.}

\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both. Defaults to
seasonally adjusted ("S").}
}
\description{
\code{ces_emp} downloads pre-packaged national employment data from the CES database.
}
\details{
This function downloads and cleans pre-packaged employment data from the CES database. The user
has four choices of employment data: non-farm payrolls, private payrolls, super sector payrolls,
or sector payrolls. The employment data is at the national level.
}
\examples{
ces_df = ces_emp(Sys.getenv("BLS_KEY"), "nfp", 2010, 2015, "U")

}
