% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ces_seriesid}
\alias{ces_seriesid}
\title{Series IDs for Current Employment Statistics (CES) data}
\usage{
ces_seriesid(adjustment, industries, data_types, states)
}
\arguments{
\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}

\item{industries}{Character vector. See vignette. List of available
industries given in \code{ces_national_codes_list} or \code{ces_state_codes_list} datasets.}

\item{data_types}{Character vector. Desired output. See series vignette. List of available
data_types given in \code{ces_national_codes_list} or \code{ces_state_codes_list} datasets.}

\item{states}{Optional character vector. See vignette. Leave blank for total US. List of
states given in \code{ces_national_codes_list} or \code{ces_state_codes_list} datasets.}
}
\value{
Vector of CES series IDs.
}
\description{
\code{ces_seriesid} constructs series IDs for downloading data from the Current Employment
Statistics (CES) database.
}
\details{
The Bureau of Labor Statistics (BLS) stores data in the form of series IDs. The structure of
series IDs varies by the underlying database (e.g., CES, JOLTS, etc.). Users must input
the correct series IDs to download their desired data. This function generates CES series IDs
for a given set of inputs.
}
\examples{
# National series for seasonally adjusted private sector employment and earnings
ces_seriesid(adjustment = c("S"), industries = c("05000000"), data_types = c("01", "03"))
}
