% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_jolts}
\alias{clean_jolts}
\title{Cleans the results of a query to the CES database.}
\usage{
clean_jolts(bls_df)
}
\arguments{
\item{bls_df}{Data frame containing results of query to the JOLTS national database.}
}
\value{
Cleaned data frame from JOLTS national database.
}
\description{
\code{clean_jolts} cleans the results of a query to the JOLTS national database and appends
additional information.
}
\details{
General data cleaning.
}
\examples{
clean_jolts(jolts_df)
}
