% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_laus}
\alias{clean_laus}
\title{Cleans the results of a query to the LAUS database.}
\usage{
clean_laus(bls_df)
}
\arguments{
\item{bls_df}{Data frame containing results of query to the LAUS national database.}
}
\value{
Cleaned data frame from LAUS national database.
}
\description{
\code{clean_laus} cleans the results of a query to the LAUS state database and appends
additional information.
}
\details{
General data cleaning.
}
\examples{
clean_laus(laus_df)
}
