% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_all_surveys}
\alias{get_all_surveys}
\title{Create and execute a query to retrieve all surveys}
\usage{
get_all_surveys(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{\link[=bls_request]{bls_request()}}}
}
\value{
a table with a survey_abbreviation and survey_name columns
}
\description{
Create and execute a query to retrieve all surveys
}
\seealso{
\code{\link{query_all_surveys}}

Other blsR-requests: 
\code{\link{bls_request}()},
\code{\link{get_latest_observation}()},
\code{\link{get_n_series_table}()},
\code{\link{get_n_series}()},
\code{\link{get_popular_series}()},
\code{\link{get_series_tables}()},
\code{\link{get_series_table}()},
\code{\link{get_series}()},
\code{\link{get_survey_info}()}
}
\concept{blsR-requests}
