% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_latest_observation}
\alias{get_latest_observation}
\title{Create and execute a query to retrieve the latest observation for a series}
\usage{
get_latest_observation(series_id, ...)
}
\arguments{
\item{series_id}{BLS series ID}

\item{...}{additional arguments to pass to \code{\link[=bls_request]{bls_request()}}}
}
\value{
a datum in the form of a list
}
\description{
Create and execute a query to retrieve the latest observation for a series
}
\seealso{
\code{\link{query_latest_observation}}

Other blsR-requests: 
\code{\link{bls_request}()},
\code{\link{get_all_surveys}()},
\code{\link{get_n_series_table}()},
\code{\link{get_n_series}()},
\code{\link{get_popular_series}()},
\code{\link{get_series_tables}()},
\code{\link{get_series_table}()},
\code{\link{get_series}()},
\code{\link{get_survey_info}()},
\code{\link{reduce_spanned_responses}()},
\code{\link{span_series_request}()}
}
\concept{blsR-requests}
