% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series_table.R
\name{get_n_series_table}
\alias{get_n_series_table}
\title{Retrieve multiple time series in one API request and return a single tibble}
\usage{
get_n_series_table(
  series_ids,
  api_key = bls_get_key(),
  start_year = NULL,
  end_year = NULL,
  year_limit = NULL,
  tidy = FALSE,
  parse_values = TRUE,
  ...
)
}
\arguments{
\item{series_ids}{a list or character vector of BLS time-series IDs. If the
items are named then the names will be used in the returned list}

\item{api_key}{Optional. An API key string. Defaults to the value returned by
\code{\link[=bls_get_key]{bls_get_key()}}. The preferred way to provide an API key is to use
\code{\link[=bls_set_key]{bls_set_key()}} or the \code{BLS_API_KEY} environment variable. Manually passing
the key will be deprecated in future releases.}

\item{start_year, end_year}{numeric 4-digit years. While optional, they are
strongly recommended. If one is provided, the other is mandatory. \code{end_year}
must be greater than \code{start_year}}

\item{year_limit}{optional number of years to paginate request by. If not
explicitly set, it will be set to 10 or 20 depending on if an \code{api_key} is
available}

\item{tidy}{optional boolean. Return will use \code{\link[=tidy_periods]{tidy_periods()}} if true}

\item{parse_values}{optional boolean. If set to \code{TRUE} (default) it will
attempt to parse the contents of \code{value} and cast numeric strings as numeric
values. If set to \code{FALSE} it will retain \code{value} as a column of strings.}

\item{...}{
  Arguments passed on to \code{\link[=get_n_series]{get_n_series}}
  \describe{
    \item{\code{series_limit}}{Maximum number of series to request in one API call
when \code{span} is set to \code{TRUE}.}
    \item{\code{span}}{when set to \code{TRUE}, requests where the number of years between
\code{start_year} and \code{end_year} exceed \code{year_limit} will be performed as
multiple requests automatically}
  }}
}
\value{
a tibble of multiple merged time series
}
\description{
Retrieve multiple time series in one API request and return a single tibble
}
\examples{
\dontrun{
get_n_series_table(
  list(uer.men ='LNS14000001', uer.women = 'LNS14000002'),
  start_year = 2005, end_year=2006
)
}

}
\seealso{
Other blsR-requests: 
\code{\link{bls_request}()},
\code{\link{get_all_surveys}()},
\code{\link{get_latest_observation}()},
\code{\link{get_n_series}()},
\code{\link{get_popular_series}()},
\code{\link{get_series_tables}()},
\code{\link{get_series_table}()},
\code{\link{get_series}()},
\code{\link{get_survey_info}()},
\code{\link{reduce_spanned_responses}()},
\code{\link{span_series_request}()}
}
\concept{blsR-requests}
