% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_popular_series}
\alias{get_popular_series}
\title{Create and execute a query to retrieve popular series}
\usage{
get_popular_series(survey_id = NA, ...)
}
\arguments{
\item{survey_id}{optional survey abbreviation}

\item{...}{additional parameters to pass to \code{\link[=bls_request]{bls_request()}}}
}
\value{
a character vector of series IDs
}
\description{
Create and execute a query to retrieve popular series
}
\seealso{
\code{\link{query_popular_series}}

Other blsR-requests: 
\code{\link{bls_request}()},
\code{\link{get_all_surveys}()},
\code{\link{get_latest_observation}()},
\code{\link{get_n_series_table}()},
\code{\link{get_n_series}()},
\code{\link{get_series_tables}()},
\code{\link{get_series_table}()},
\code{\link{get_series}()},
\code{\link{get_survey_info}()}
}
\concept{blsR-requests}
