% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_survey_info}
\alias{get_survey_info}
\title{Create and execute a query to retrieve information about a survey}
\usage{
get_survey_info(survey_id, ...)
}
\arguments{
\item{survey_id}{survey abbreviation}

\item{...}{additional parameters to pass to \link{bls_request}}
}
\value{
a list of survey information
}
\description{
Create and execute a query to retrieve information about a survey
}
\seealso{
\code{\link{query_survey_info}}

Other blsR-requests: 
\code{\link{bls_request}()},
\code{\link{get_all_surveys}()},
\code{\link{get_latest_observation}()},
\code{\link{get_n_series_table}()},
\code{\link{get_n_series}()},
\code{\link{get_popular_series}()},
\code{\link{get_series_tables}()},
\code{\link{get_series_table}()},
\code{\link{get_series}()}
}
\concept{blsR-requests}
