% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jolts.R
\name{jolts_layoffs}
\alias{jolts_layoffs}
\title{Download layoffs data from the Job Opening and Labor Turnover survey (JOLTS).}
\usage{
jolts_layoffs(
  bls_key,
  series = "nfp",
  start_year,
  end_year,
  adjustment,
  data_levels = "R"
)
}
\arguments{
\item{bls_key}{BLS API key for the user. See vignette.}

\item{series}{Character. Data series. Either non-farm layoffs ("nfp"), private layoffs
("private"), supersector layoffs ("super"), or sector layoffs ("sector"). Defaults to non-farm
layoffs.}

\item{start_year}{Numeric. Year to start data download.}

\item{end_year}{Numeric. Year to end data download.}

\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}

\item{data_levels}{String. Rate ("R") or level ("L"). Defaults to rate as reported by JOLTS.}
}
\description{
\code{jolts_seps} downloads pre-packaged layoffs data from the JOLTS database.
}
\details{
This function downloads and cleans pre-packaged layoffs data from the JOLTS database. The
user has three choices for layoffs data: non-farm layoffs, private layoffs, or super
sector hiring. The data is formatted in terms of rates by default to be similar to the quoted
JOLTS series. The function reflects the October 2020 update to the JOLTS data series.
}
\examples{
jolts_df = jolts_layoffs(Sys.getenv("BLS_KEY"), "nfp", 2010, 2015, "U")

}
