% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jolts.R
\name{jolts_seps}
\alias{jolts_seps}
\title{Download separations data from the Job Opening and Labor Turnover survey (JOLTS).}
\usage{
jolts_seps(bls_key, series = "nfp", start_year, end_year, adjustment)
}
\arguments{
\item{bls_key}{BLS API key for the user. See vignette.}

\item{series}{Character. Data series. Either non-farm separations ("nfp"), private separations
("private"), supersector separations ("super"), or sector separations ("sector"). Defaults to
non-farm separations.}

\item{start_year}{Numeric. Year to start data download.}

\item{end_year}{Numeric. Year to end data download.}

\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}
}
\description{
\code{jolts_seps} downloads pre-packaged separations data from the JOLTS database.
}
\details{
This function downloads and cleans pre-packaged separations data from the JOLTS database. The
user has three choices for separations data: non-farm separations, private separations, or super
sector hiring. The data is formatted in terms of rates similar to the quoted JOLTS
series. Separations is equal to the sum of layoffs, quits, and other separations. The function
reflects the October 2020 update to the JOLTS data series.
}
\examples{
jolts_df = jolts_seps(Sys.getenv("BLS_KEY"), "nfp", 2010, 2015, "U")

}
