% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jolts.R
\name{jolts_seriesid}
\alias{jolts_seriesid}
\title{Series IDs for Job Openings and Labor Turnover Survey (JOLTS) data}
\usage{
jolts_seriesid(
  adjustment,
  industries,
  data_types,
  data_levels,
  states,
  areas,
  sizes
)
}
\arguments{
\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}

\item{industries}{Character vector. See vignette. List of available
industries given in \code{jolts_codes_list} dataset.}

\item{data_types}{Character vector. Desired output. See series vignette. List of available
data_types given in \code{jolts_codes_list} dataset.}

\item{data_levels}{Character vector. Levels ("L") or rates ("R") or both.}

\item{states}{Option character vector. See vignette. Leave blank for total US. List of
states given in \code{jolts_codes_list} dataset.}

\item{areas}{Optional character vector. See vignette. Leave blank for total US. List of
areas in \code{jolts_codes_list} dataset.
#' @param sizes Optional character vector. See vignette. Leave blank for all. List
in \code{jolts_codes_list} dataset.}
}
\value{
Vector of JOLTS series IDs.
}
\description{
\code{jolts_seriesid} constructs series IDs for downloading data from the Job Openings and Labor
Turnover Survey (JOLTS) database.
}
\details{
The Bureau of Labor Statistics (BLS) stores data in the form of series IDs. The structure of
series IDs varies by the underlying database (e.g., CES, JOLTS, etc.). Users must input
the correct series IDs to download their desired data. This function helps overcome this problem
by automatically generating JOLTS series IDs for a given set of inputs. Note that only total
non-farm data series IDs can be extracted for different regions of the United States. Reflects
the October 2020 updates to JOLTS data series.
}
\examples{
# National series for seasonally adjusted private sector hires and job openings
jolts_seriesid(adjustment = c("S"), industries = c("100000"), data_types = c("HI", "JO"),
data_levels = c("L"))
}
