% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus.R
\name{laus_seriesid}
\alias{laus_seriesid}
\title{Series IDs for Local Area Unemployment Statistics (LAUS) data}
\usage{
laus_seriesid(adjustment, states, data_types)
}
\arguments{
\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}

\item{states}{Character vector. See vignette and \code{laus_codes_list} object.}

\item{data_types}{Character vector. See vignette and \code{laus_codes_list} object.}
}
\value{
Vector of LAUS series IDs.
}
\description{
\code{laus_seriesid} constructs series IDs for downloading data from the Local Area
Unemployment Statistics (LAUS) database.
}
\details{
The Bureau of Labor Statistics (BLS) stores data in the form of series IDs. The structure of
series IDs varies by the underlying database (e.g., CES, JOLTS, LAUS, etc.). Users must input
the correct series IDs to download their desired data. This function generates LAUS series IDs
given a set of user inputs.
}
\examples{
laus_seriesid(adjustment = "U", states = c("ST0100000000000", "ST0200000000000"), data_types = c("03"))
}
