% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus.R
\name{laus_unemp}
\alias{laus_unemp}
\title{Download state unemployment numbers from the Local Area Unemployment Statistics (LAUS) database}
\usage{
laus_unemp(bls_key, states = "all", start_year, end_year, adjustment)
}
\arguments{
\item{bls_key}{BLS API key for the user. See vignette.
#' @param states Character vector. All states ("all"), Washington D.C. ("DC"), or list of state
codes. See vignette and \code{laus_codes_list} object.}

\item{start_year}{Year to start data download.}

\item{end_year}{Year to end data download.}

\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}
}
\value{
Cleaned data frame of LAUS data.
}
\description{
\code{laus_unemp} downloads state unemployment numbers from the Local Area Unemployment
Statistics (LAUS) database.
}
\details{
This function downloads and cleans pre-packaged state unemployment levels data from the LAUS
database. The default setting is to download unemployment levels for all states in the U.S.
excluding D.C. The user can shorten or lengthen the list of states but is constrained by the
BLS 50 series limit.
}
\examples{
laus_unemp(Sys.getenv("BLS_KEY"), "all", 2010, 2015, "U")
}
