% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tables.R
\name{merge_tidy_tables}
\alias{merge_tidy_tables}
\title{Turn a list of one or more series into a single table of time series data}
\usage{
merge_tidy_tables(tidy_tables)
}
\arguments{
\item{tidy_tables}{a named list of tables with matching periodicity. Mixing
data with different (monthly, quarterly, annual) periodicity is unsupported.
The list names will be used as column names in the output.}
}
\value{
tibble
}
\description{
\code{merge_tidy_tables()} turns a list of series as returned by
\code{\link[=data_as_tidy_table]{data_as_tidy_table()}} into a single tibble
}
\seealso{
Other blsR-utils: 
\code{\link{bls-api-key}},
\code{\link{data_as_table}()},
\code{\link{data_as_tidy_table}()},
\code{\link{merge_tables}()},
\code{\link{reduce_spanned_responses}()},
\code{\link{span_request_queries}()},
\code{\link{span_series_request}()},
\code{\link{tidy_periods}()},
\code{\link{tidy_table_as_zoo}()}
}
\concept{blsR-utils}
