% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{okay_ces_seriesid}
\alias{okay_ces_seriesid}
\title{Check whether CES series ID is valid.}
\usage{
okay_ces_seriesid(adjustment, industries, data_types, states)
}
\arguments{
\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}

\item{industries}{Character vector. See vignette. List of available
industries given in \code{ces_national_codes_list} or \code{ces_state_codes_list} datasets.}

\item{data_types}{Character vector. Desired output. See series vignette. List of available
data_types given in \code{ces_national_codes_list} or \code{ces_state_codes_list} datasets.}

\item{states}{Optional character vector. See vignette. Leave blank for total US. List of
states given in \code{ces_national_codes_list} or \code{ces_state_codes_list} datasets.}
}
\value{
TRUE if the series IDs are valid. FALSE otherwise.
}
\description{
\code{okay_ces_seriesid} checks whether the inputs yield a valid CES series ID string.
}
\details{
General error checking for CES data downloads.
}
\examples{
okay_ces_seriesid("S", "0000000", "01")
}
