% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus.R
\name{okay_laus_seriesid}
\alias{okay_laus_seriesid}
\title{Check whether LAUS series ID is valid.}
\usage{
okay_laus_seriesid(adjustment, states, data_types)
}
\arguments{
\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}

\item{states}{Character vector. See vignette and \code{laus_codes_list} object.}

\item{data_types}{Character vector. See vignette and \code{laus_codes_list} object.}
}
\value{
TRUE if the series IDs are valid. FALSE otherwise.
}
\description{
\code{okay_laus_seriesid} checks whether the given inputs yield a valid LAUS series ID string.
}
\details{
General error checking for LAUS data downloads.
}
\examples{
okay_laus_seriesid(adjustment = "U", states = c("ST0100000000000"), data_types = c("03"))
}
