% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{okay_series_input}
\alias{okay_series_input}
\title{Checks series ID inputs.}
\usage{
okay_series_input(inputs, valid_inputs)
}
\arguments{
\item{inputs}{Character vector. List of inputs to series ID string.}

\item{valid_inputs}{Character vector. List of valid inputs to series ID string.}
}
\value{
TRUE if the series inputs are valid. FALSE otherwise.
}
\description{
\code{okay_series_input} returns whether a series ID input is valid.
}
