% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_all_surveys}
\alias{query_all_surveys}
\title{Create a query to retrieve all surveys}
\usage{
query_all_surveys()
}
\value{
list of query parameters
}
\description{
Create a query to retrieve all surveys
}
\seealso{
Other blsR-queries: 
\code{\link{query_latest_observation}()},
\code{\link{query_n_series}()},
\code{\link{query_popular_series}()},
\code{\link{query_series}()},
\code{\link{query_survey_info}()},
\code{\link{span_request_queries}()}
}
\concept{blsR-queries}
