% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_latest_observation}
\alias{query_latest_observation}
\title{Create a Query to retrieve the latest observation for a time series}
\usage{
query_latest_observation(series_id)
}
\arguments{
\item{series_id}{BLS series ID}
}
\value{
list of query parameters
}
\description{
Create a Query to retrieve the latest observation for a time series
}
\seealso{
Other blsR-queries: 
\code{\link{query_all_surveys}()},
\code{\link{query_n_series}()},
\code{\link{query_popular_series}()},
\code{\link{query_series}()},
\code{\link{query_survey_info}()}
}
\concept{blsR-queries}
