% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_popular_series}
\alias{query_popular_series}
\title{Create a query to retrieve popular series}
\usage{
query_popular_series(survey_id = NA)
}
\arguments{
\item{survey_id}{string optional}
}
\value{
list of query parameters
}
\description{
Create a query to retrieve popular series
}
\examples{
popular_series_query <- query_popular_series()
popular_labor_force_series <- query_popular_series('LN')

}
\seealso{
Other blsR-queries: 
\code{\link{query_all_surveys}()},
\code{\link{query_latest_observation}()},
\code{\link{query_n_series}()},
\code{\link{query_series}()},
\code{\link{query_survey_info}()}
}
\concept{blsR-queries}
