% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces.R
\name{ces_emp}
\alias{ces_emp}
\title{Download employment data from Current Employment Statistics (CES) survey}
\usage{
ces_emp(bls_key, series = "nfp", start_year, end_year, adjustment)
}
\arguments{
\item{bls_key}{BLS API key for the user. See vignette.}

\item{series}{Character. Data series. Either non-farm payrolls ("nfp"), private payrolls
("private"), or supersector payrolls ("super"). Defaults to non-farm payrolls.}

\item{start_year}{Numeric. Year to start data download.}

\item{end_year}{Numeric. Year to end data download.}

\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}
}
\description{
\code{ces_emp} downloads pre-packaged employment data from the CES database.
}
\details{
This function downloads and cleans pre-packaged employment data from the CES database. The user
has three choices of employment data: non-farm payrolls, private payrolls, or super sector
payrolls.
}
\examples{
ces_df = ces_emp(Sys.getenv("BLS_KEY"), "nfp", 2010, 2015, "U")

}
