% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jolts.R
\name{jolts_download}
\alias{jolts_download}
\title{Download data from the Job Opening and Labor Turnover survey (JOLTS)}
\usage{
jolts_download(bls_key, start_year, end_year, adjustment, industries,
  data_types, data_levels, regions, clean = TRUE)
}
\arguments{
\item{bls_key}{BLS API key for the user. See vignette.}

\item{start_year}{Year to start data download.}

\item{end_year}{Year to end data download.}

\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}

\item{industries}{Character vector. See vignette. List of available
industries given in \code{jolts_codes_list} or \code{jolts_codes_list} datasets.}

\item{data_types}{Character vector. Desired output. See series vignette. List of available
data_types given in \code{jolts_codes_list} or \code{jolts_codes_list} datasets.}

\item{data_levels}{Character vector. Rates ("R) or Levels ("L") or both.}

\item{regions}{Optional character vector. See vignette. Leave blank for total US. List of
regions given in \code{jolts_codes_list} or \code{jolts_codes_list} datasets.}

\item{clean}{Optional logical. Whether to clean the data into usable format or not.
Defaults to TRUE.}
}
\value{
Cleaned data frame of JOLTS data.
}
\description{
\code{jolts_download} retrieves and downloads data from the Job Opening and Labor Turnover
survey (JOLTS) database.
}
\details{
This function constructs BLS series IDs for a chosen set of parameters and then downloads
the data from the JOLTS database. This function also cleans the data and appends on identifiers
for easy panel formatting.
}
\examples{
Add in from our other exercise.

}
