% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jolts.R
\name{jolts_openings}
\alias{jolts_openings}
\title{Download job openings data from the Job Opening and Labor Turnover survey (JOLTS).}
\usage{
jolts_openings(
  bls_key,
  series = "nfp",
  start_year,
  end_year,
  adjustment,
  data_levels = "L"
)
}
\arguments{
\item{bls_key}{BLS API key for the user. See vignette.}

\item{series}{Character. Data series. Either non-farm openings ("nfp"), private openings
("private"), supersector openings ("super"), or sector openings ("sector"). Defaults to non-farm
openings.}

\item{start_year}{Numeric. Year to start data download.}

\item{end_year}{Numeric. Year to end data download.}

\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}

\item{data_levels}{String. Rate ("R") or level ("L"). Defaults to level as reported by JOLTS.}
}
\description{
\code{jolts_openings} downloads pre-packaged openings data from the JOLTS database.
}
\details{
This function downloads and cleans pre-packaged job openings data from the JOLTS database. The
user has three choices for openings data: non-farm openings, private openings, or super
sector hiring. The data is formatted in terms of levels by default to be similar to the quoted
JOLTS series. The function reflects the October 2020 update to the JOLTS data series.
}
\examples{
jolts_df = jolts_openings(Sys.getenv("BLS_KEY"), "nfp", 2010, 2015, "U")

}
