% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jolts.R
\name{okay_jolts_seriesid}
\alias{okay_jolts_seriesid}
\title{Check whether JOLTS series ID is valid.}
\usage{
okay_jolts_seriesid(adjustment, industries, data_types, data_levels,
  regions)
}
\arguments{
\item{adjustment}{Character vector. Seasonal adjustment ("S") or not ("U") or both.}

\item{industries}{Character vector. See vignette. List of available
industries given in \code{jolts_codes_list} or \code{jolts_codes_list} datasets.}

\item{data_types}{Character vector. Desired output. See series vignette. List of available
data_types given in \code{jolts_codes_list} or \code{jolts_codes_list} datasets.}

\item{data_levels}{Character vector. Rates ("R) or Levels ("L") or both.}

\item{regions}{Optional character vector. See vignette. Leave blank for total US. List of
regions given in \code{jolts_codes_list} or \code{jolts_codes_list} datasets.}
}
\value{
TRUE if the series IDs are valid. FALSE otherwise.
}
\description{
\code{okay_jolts_seriesid} checks whether the inputs yield a valid JOLTS series ID string.
}
\details{
General error checking for JOLTS data downloads.
}
\examples{
NA
}
